requirements.txt
opencv-python==4.4.0.44
PyWavelets>=1.1.1
matplotlib==3.6.0
natsort==8.1.0
numpy==1.23.2
pandas==1.4.2
Pillow==9.2.0
scikit_learn==1.1.2
scipy==1.9.1
tensorflow==2.10.0
tqdm==4.64.1
